requirements.txt
torch>=2.2.0,<2.7
torchaudio>=2.2.0,<2.7
tokenizers
transformers
huggingface_hub
moshi==0.2.2
torchtune
torchao
silentcipher @ git+https://github.com/SesameAILabs/silentcipher@master
flask
werkzeug
SpeechRecognition
PyAudio
sacremoses
gTTS
pydub
ffmpeg-python
# Nouvelles dépendances pour la version LLM
mistralai>=0.4.0
python-dotenv>=1.0.0
deep-translator>=1.11.4